feat: deployment configuration and frontend refactoring
- Added deployment configs for Fly.io (backend) and Vercel (frontend)
- Added CI/CD workflows in .github
- Refactored frontend components to use custom hooks (useChatStream, useResizable)
- Extracted interpolation logic to shared utility
- Updated README with deployment instructions
