+++ .gitignore
@@ -69,3 +69,4 @@ htmlcov/
 # Misc
 *.bak
 .cache/
+.vercel
