+++ backend/fly.toml
@@ -1,0 +1,21 @@
+app = "prompt-studio-api"
+primary_region = "iad"
+
+[build]
+  dockerfile = "Dockerfile"
+
+[env]
+  PORT = "8080"
+
+[http_service]
+  internal_port = 8080
+  force_https = true
+  auto_stop_machines = "off"
+  auto_start_machines = true
+  min_machines_running = 1
+
+[deploy]
+  # Run migrations and seed provider content before each deploy
+  release_command = "uv run alembic upgrade head && uv run python scripts/seed_provider_content.py"
+
+
