+++ frontend/package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "dev": "vite",
     "build": "tsc -b && vite build",
-    "preview": "vite preview --open"
+    "preview": "vite preview --open",
+    "lint": "tsc --noEmit"
   },
   "dependencies": {
     "highlight.js": "^11.11.1",
